refactor: Put Task related classes into package dude.task
