+++ src/main/java/dude/Dude.java
@@ -8,6 +8,12 @@ import dude.exception.DudeNullDateTimeException;
 import dude.exception.DudeNullDescriptionException;
 import dude.exception.DudeNumberException;
 
+import dude.task.Deadline;
+import dude.task.Event;
+import dude.task.Task;
+import dude.task.TaskList;
+import dude.task.ToDo;
+
 import java.time.LocalDateTime;
 
 /**
