+++ src/main/java/dude/task/Deadline.java
@@ -1,4 +1,4 @@
-package dude;
+package dude.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
