+++ src/main/java/dude/task/Event.java
@@ -1,4 +1,4 @@
-package dude;
+package dude.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
