+++ src/main/java/dude/task/TaskList.java
@@ -1,4 +1,4 @@
-package dude;
+package dude.task;
 
 import dude.exception.DudeDuplicatedTaskException;
